replace base64 with data-encoding and move utoipa related stuff into a separate file
