+++ Cargo.toml
@@ -9,7 +9,16 @@ edition = "2021"
 base64 = "^0.21.5"
 bincode = "^1"
 bs58 = "^0.5.0"
-derive_more = { version = "^0.99", default-features = false, features = ["as_ref", "deref", "from", "into"] }
+data-encoding = "^2.5.0"
+derive_more = { version = "^0.99", default-features = false, features = [
+  "as_mut",
+  "as_ref",
+  "deref",
+  "deref_mut",
+  "from",
+  "into",
+  "into_iterator",
+] }
 serde = "^1"
 solana-program = "~1.17"
 solana-sdk = "~1.17"
