+++ src/lib.rs
@@ -4,6 +4,7 @@ pub use b64_buffer::*;
 pub use b64_legacy_tx::*;
 pub use b64_versioned_tx::*;
 pub use u64_str::*;
+pub use utoipa::*;
 
 pub mod b58_pubkey;
 pub mod b58_signature;
@@ -11,3 +12,4 @@ pub mod b64_buffer;
 pub mod b64_legacy_tx;
 pub mod b64_versioned_tx;
 pub mod u64_str;
+pub mod utoipa;
